feat: implement request constructor
