+++ src/types/request.rs
@@ -1,5 +1,6 @@
 use std::net::*;
 use std::io::{Write};
+use std::io::{Read};
 
 pub struct Request {
   pub url: String,
@@ -16,5 +17,22 @@ impl Request {
     }
     // Close connection
     self.socket.shutdown(Shutdown::Both).expect("shutdown call failed");
+  }
+
+  pub fn from(mut stream: TcpStream) -> Request {
+    let mut buf = [0u8 ;4096];
+    match stream.read(&mut buf) {
+        Ok(_) => {
+            let req_str = String::from_utf8_lossy(&buf);
+            let req_parts = req_str.split(" "); 
+            let vec: Vec<&str> = req_parts.collect();
+            Request {
+                url: String::from(vec[1]),
+                req_type: String::from(vec[0]),
+                socket: stream
+            }
+            },
+        Err(e) => panic!("Error: {:?}", e),
+    }
   }
 }
